Change no init.fish fallback behaviour
This commit changes how fundle treats plugins with no `init.fish` files.

The previous behaviour sourced all the fish files in the root folder of
the package, which allowed compatibility with fish themes, but it did so
by sourcing files defining inline functions, as if they were typed
directly in the shell. The side effect of this is that user defined
functions overriding that behaviour wouldn't work.

The new behavior includes the root folder of the plugin in the
`fish_function_path` variable, while preserving the higher precedence of
user defined functions. This way, the user is able to override theme
functions such as `fish_greeting`.