feat: Add support for @<tag> in plugin name
Choosing a specific release of a plugin should be done in a user
friendly way. Therefore, it should be as simple as adding a tag
at the end of the plugin name consisting of username/repository.
This small modifications makes this possible using '@' as a delimiter.

Example:
    `fundle plugin 'joseluisq/gitnow@2.7.0'`

Not selecting a tag still acts the same as usual.